% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bal.fe.psa.R
\name{bal.fe.psa}
\alias{bal.fe.psa}
\title{Fisher's Exact Test for Independence of Treatments within Strata}
\usage{
bal.fe.psa(categorical, treatment = NULL, strata = NULL, FB = 2000)
}
\arguments{
\item{categorical}{Categorical covariate that is being balanced within
strata in a PSA. If \code{categorical} has three columns, then the second
and third are assumed to be the treatment and strata respectively.  Missing
values are not allowed. May be factor or numeric.}

\item{treatment}{Binary variable of same length as \code{categorical};
generally 0 for 'control,' 1 for 'treatment.'}

\item{strata}{Integer variable; a vector of same length as
\code{categorical} indicating the derived strata from estimated propensity
scores.}

\item{FB}{Numeric; number of replications sent to fisher.test.}
}
\value{
Returns list of the same lenght as the number of strata containing
p-values for the indpendence of treatment within each stratum derived from
Fisher's Exact test.
}
\description{
Simple function that calls fisher.test repeatedly for each strata, testing
the independence of treatements for the given covariate within strata.
}
\details{
This function makes repeated calls to fisher.test, Fisher's Exact test, to
test whether the distribution of the covariate categorical is independent of
treatment within each stratum; a list of p-values for the test for each
stratum are returned.
}
\examples{
#Everything random
categorical<-sample(4, 1000, replace = TRUE)
treatment<-sample(c(0,1), 1000, replace = TRUE)
strata<-sample(5, 1000, replace = TRUE)
bal.fe.psa(categorical, treatment, strata)

#Perfect balance on 80\%, random on last 20\%
categorical<-rep(sample(5,1000, replace=TRUE), 2)
treatment<-c(rep(0,1000), rep(1,1000))
strata<-sample(6, 1200, replace=TRUE)
strata<-c(strata[1:1000], strata[1:800], strata[1001:1200])
bal.fe.psa(categorical, treatment, strata)

}
\seealso{
\code{bal.cs.psa}, \code{bal.ms.psa}, \code{bal.ks.psa}
}
\author{
James E. Helmreich \email{ James.Helmreich@Marist.edu}

Robert M. Pruzek \email{RMPruzek@yahoo.com}
}
\keyword{htest}
