% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSAgraphics-package.R
\docType{package}
\name{granova-package}
\alias{granova-package}
\alias{PSAgraphics-package}
\alias{PSAgraphics}
\title{Graphical Analysis of Variance}
\description{
A collection of functions that primarily produce graphics to aid in a
Propensity Score Analysis (PSA).  Functions include: cat.psa and box.psa to
test balance within strata of categorical and quantitative covariates,
circ.psa for a representation of the estimated effect size by stratum,
loess.psa that provides a graphic and loess based effect size estimate, and
various balance functions that provide measures of the balance achieved via
a PSA in a categorical covariate.

A collection of functions that primarily produce graphics to aid in a
Propensity Score Analysis (PSA).  Functions include: cat.psa and box.psa to
test balance within strata of categorical and quantitative covariates,
circ.psa for a representation of the estimated effect size by stratum,
loess.psa that provides a graphic and loess based effect size estimate, and
various balance functions that provide measures of the balance achieved via
a PSA in a categorical covariate.
}
\details{
\tabular{ll}{ Package: \tab PSAgraphics\cr Version: \tab 2.0\cr License:
\tab GPL (>= 2)\cr }

\tabular{ll}{ Package: \tab PSAgraphics\cr Version: \tab 2.0\cr License:
\tab GPL (>= 2)\cr }
}
\seealso{
\code{\link{box.psa}} \code{\link{cat.psa}} \code{\link{circ.psa}}
\code{\link{loess.psa}} \code{\link{bal.ks.psa}} \code{\link{bal.ms.psa}}

\code{\link{bal.fe.psa}} \code{\link{cstrata.psa}}
\code{\link{cv.trans.psa}} \code{\link{cv.bal.psa}}

\code{\link{box.psa}} \code{\link{cat.psa}} \code{\link{circ.psa}}
\code{\link{loess.psa}} \code{\link{bal.ks.psa}} \code{\link{bal.ms.psa}}

\code{\link{bal.fe.psa}} \code{\link{cstrata.psa}}
\code{\link{cv.trans.psa}} \code{\link{cv.bal.psa}}
}
\author{
James E. Helmreich <James.Helmreich@Marist.edu> and Robert M. Pruzek <RMPruzek@yahoo.com>

Maintainer: Jason Bryer <jason@bryer.org>

James E. Helmreich <James.Helmreich@Marist.edu> and

Robert M. Pruzek <RMPruzek@yahoo.com>

Maintainer: James E. Helmreich <James.Helmreich@Marist.edu>
}
\keyword{hplot}
